VMO-4136 [Validation API] Validation should delegate to individual block stores to each validate themselves Part 1
