+++ src/store/flow/block-types/Core_OutputBlockStore.ts
@@ -4,6 +4,7 @@ import {IBlock, IBlockExit} from '@floip/flow-runner'
 import {IdGeneratorUuidV4} from '@floip/flow-runner/dist/domain/IdGeneratorUuidV4'
 import {IOutputBlock} from '@floip/flow-runner/src/model/block/IOutputBlock'
 import {defaultsDeep} from 'lodash'
+import {validate_community_block} from '@/store/validation/validationHelpers'
 import {IFlowsState} from '../index'
 
 export const BLOCK_TYPE = 'Core.Output'
@@ -45,6 +46,10 @@ export const actions: ActionTree<IFlowsState, IRootState> = {
       blockId: block.uuid,
       test: 'block.value > 0',
     }, {root: true})
+  },
+
+  validate_block({rootGetters}, {block, schemaVersion}: {block: IBlock, schemaVersion: string}) {
+    return validate_community_block({block, schemaVersion})
   },
 }
