+++ src/store/flow/block-types/Core_SetGroupMembershipStore.ts
@@ -3,6 +3,7 @@ import {IRootState} from '@/store'
 import {IBlock, IBlockExit, ISetGroupMembershipBlockConfig} from '@floip/flow-runner'
 import {IdGeneratorUuidV4} from '@floip/flow-runner/dist/domain/IdGeneratorUuidV4'
 import {defaultsDeep} from 'lodash'
+import {validate_community_block} from '@/store/validation/validationHelpers'
 import {IFlowsState} from '../index'
 
 export interface IGroupOption {
@@ -64,6 +65,10 @@ export const actions: ActionTree<IFlowsState, IRootState> = {
       path: 'is_member',
       value: isMember,
     }, {root: true})
+  },
+
+  validate_block({rootGetters}, {block, schemaVersion}: {block: IBlock, schemaVersion: string}) {
+    return validate_community_block({block, schemaVersion})
   },
 }
