+++ src/store/flow/block-types/MobilePrimitives_SelectManyResponseBlockStore.ts
@@ -12,6 +12,7 @@ import {
   stateFactory,
 } from './MobilePrimitives_SelectOneResponseBlockStore'
 import {IBlockWithBranchingType, OutputBranchingType} from '@/components/interaction-designer/block-editors/BlockOutputBranchingConfig.vue'
+import {validate_community_block} from '@/store/validation/validationHelpers'
 
 export const BLOCK_TYPE = 'MobilePrimitives.SelectManyResponse'
 
@@ -89,6 +90,10 @@ export const actions: ActionTree<ICustomFlowState, IRootState> = {
       vendor_metadata: {},
       tags: [],
     })
+  },
+
+  validate_block({rootGetters}, {block, schemaVersion}: {block: IBlock, schemaVersion: string}) {
+    return validate_community_block({block, schemaVersion})
   },
 }
